Move remaining screen functions
